% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add_norway_map}
\alias{add_norway_map}
\title{Add a map of Norway ontop of a ggplot object}
\usage{
add_norway_map(crs = NULL, bbox = NULL, ...)
}
\description{
Add a map of Norway ontop of a ggplot object
}
