% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_statistics.R
\name{empirical_chi}
\alias{empirical_chi}
\title{This function takes in a data object that has been created by the
extract_extreme_fields() function. Then, a sliding window approach
is performed over the distances to the conditioning sites,
for computing the empirical exceedance probabilities given a threshold τ, for all observations
with a distance d ± Δd to a conditioning site, for a range of different values of d and τ.}
\usage{
empirical_chi(
  data,
  thresholds,
  dist_centers,
  dist_radius = 0.5,
  add_zero_dist = TRUE
)
}
\description{
The input variables are:
data: This is a data object created by the extract_extreme_fields() function
threshold: A vector of thresholds used for computing empirical exceedance probabilities
dist_centers: The values of d used for computing empirical exceedance probabilities
dist_radius: The value of Δd
add_zero_dist: Boolean variable. Should we add rows for d = 0 in the output?
}
\details{
The output of the function is a data.frame with one row for each combination of
the given values of d and τ. The data.frame has columns:
chi: Empirical exceedance probability, for the threshold τ, for all observations with
distance d ± Δd to a conditioning site of interest.
n: Number of observations with distance d ± Δd to a conditioning site that exceeds a
threshold of τ
Threshold: Value of τ.
dist: Value of d.
}
