% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_statistics.R
\name{empirical_moments}
\alias{empirical_moments}
\title{This function takes in a data object that has been created by the
extract_extreme_fields() function. Then, a sliding window approach
is performed over the distances to the conditioning sites and the
values of the threshold exceedances at the conditioning sites,
for computing the empirical mean and standard deviation of all observations
with a distance d ± Δd to a conditioning site where a threshold exceedance log(y0) ± Δy0 was
observed, for a range of different values of d and y0.}
\usage{
empirical_moments(
  data,
  dist_centers,
  log_y0_centers,
  dist_radius = 0.5,
  log_y0_radius = 0.05,
  add_zero_dist = TRUE
)
}
\description{
The input variables are:
data: This is a data object created by the extract_extreme_fields() function
dist_centers: The values of d used for computing empirical moments
log_y0_centers: The values of log(y0) used for computing empirical moments
dist_radius: The value of Δd
log_y0_radius: The value of Δy0
add_zero_dist: Boolean variable. Should we add rows for d = 0 in the output?
}
\details{
The output of the function is a data.frame with one row for each combination of
the given values of d and log(y0). The data.frame has columns:
mean: Empirical mean of all observations with distance d ± Δd to a conditioning site where a
threshold exceedance log(y0) ± Δy0 was observed.
sd: Empirical standard deviation of all observations with distance d ± Δd to a conditioning site
where a threshold exceedance log(y0) ± Δy0 was observed.
log_y0: Value of log(y0).
dist: Value of d.
n: Number of observations distance d ± Δd to a conditioning site where a
threshold exceedance log(y0) ± Δy0 was observed.
}
