% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{get_s0_index}
\alias{get_s0_index}
\title{This function takes in a (d x 2)-dimensional matrix of coordinates from a regular grid
together with a distance Δs0, and then it returns the indices of a subgrid of the
original grid, that has a resolution of Δs0 x Δs0}
\usage{
get_s0_index(coords, delta_s0 = 1)
}
\description{
This function takes in a (d x 2)-dimensional matrix of coordinates from a regular grid
together with a distance Δs0, and then it returns the indices of a subgrid of the
original grid, that has a resolution of Δs0 x Δs0
}
