% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{locate_nc_files_in_metno_page}
\alias{locate_nc_files_in_metno_page}
\title{Given the url to a thredds.met.no webpage,
locate all files in the webpage that satisfy certain requirements}
\usage{
locate_nc_files_in_metno_page(url)
}
\description{
Given the url to a thredds.met.no webpage,
locate all files in the webpage that satisfy certain requirements
}
