% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-distributions.R
\name{pgp}
\alias{pgp}
\title{Cumulative distribution function for the generalised Pareto (GP) distribution}
\usage{
pgp(x, s, xi, u = 0, alpha = 0.5)
}
\description{
Input variables:
x: The input x to the function F(x)
s: The scale parameter of the GP distribution
xi: The shape parameter of the GP distribution
u: The location parameter of the GP distribution
alpha: a probability such that the scale parameter s is equal to the
alpha-quantile of the GP distribution
}
