% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-distributions.R
\name{plaplace}
\alias{plaplace}
\title{Cumulative distribution function for the Laplace distribution}
\usage{
plaplace(x)
}
\description{
Cumulative distribution function for the Laplace distribution
}
