% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-distributions.R
\name{qgp}
\alias{qgp}
\title{Quantile function for the generalised Pareto (GP) distribution}
\usage{
qgp(p, s, xi, u = 0, alpha = 0.5)
}
\description{
Input variables:
p: The input p to the function Q(p)
s: The scale parameter of the GP distribution
xi: The shape parameter of the GP distribution
u: The location parameter of the GP distribution
alpha: a probability such that the scale parameter s is equal to the
alpha-quantile of the GP distribution
}
