% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{simulate_intensity_posterior}
\alias{simulate_intensity_posterior}
\title{Simulate spatial realisations of the precipitation intensity process.}
\usage{
simulate_intensity_posterior(
  samples,
  threshold,
  coords,
  s0_index,
  spde,
  get_a_func,
  get_b_func,
  get_Q,
  get_tau,
  verbose = FALSE,
  n_cores = 1,
  n_per_sample = 1
)
}
\description{
The input variables are:
samples: an (n x m)-dimensional matrix of hyperparameter samples from the fitted
spatial conditional extremes model, where n is the total number of samples and
m is the number of hyperparameters in our chosen model
threshold: The threshold of the spatial conditional extremes model
coords: a (d x 2)-dimensional matrix of the coordinates for all the d locations in which
we want to simulate data
s0_index: an integer between 1 and d, describing which of the d elements in coords that
correspond to the conditioning site
spde: an INLA SPDE object which we use for sampling from the SPDE approximation
get_a_func: a function that takes in an m-dimensional vector of hyperparameters and returns
a function that computes the standardising function a() of the spatial conditional
extremes model
get_b_func: a function that takes in an m-dimensional vector of hyperparameters and returns
a function that computes the standardising function b() of the spatial conditional
extremes model
get_Q: a function that takes in an m-dimensional vector of hyperparameters and returns
a function that computes the precision matrix of the SPDE approximation
get_tau: a function that takes in an m-dimensional vector of hyperparameters and returns
a function that computes the precision of the nugget effect
verbose: a Boolean: Should we be verbose or not?
n_cores: an integer describing the number of parallel cores that should be used for
running the function
n_per_sample: number of spatial realisations of the intensity process that we will
simulate for each of the n hyperparameter samples
}
