% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{simulate_probit_posterior}
\alias{simulate_probit_posterior}
\title{Simulate spatial realisations of the precipitation occurrence process using
the non-spatial probit model}
\usage{
simulate_probit_posterior(
  fixed_samples,
  coords,
  s0_index,
  create_X,
  replace_zeros_at_s0 = FALSE,
  verbose = FALSE,
  n_cores = 1,
  n_per_sample = 1
)
}
\description{
The input variables are:
fixed_samples: an (n x k)-dimensional matrix of samples for the parameters of the fixed
effects of the spatial probit model, i.e., those describing the spline function in
the mean structure of the latent field
coords: a (d x 2)-dimensional matrix of the coordinates for all the d locations in which
we want to simulate data
s0_index: an integer between 1 and d, describing which of the d elements in coords that
correspond to the conditioning site
create_X: a function that takes in the distance to the conditioning site and outputs
a design matrix with all the basis functions for the spline function
replace_zeros_at_s0: a Boolean: Should we enforce I(s0) = 1 by resampling from the model
whenever we get a realisation that gives I(s0) = 0?
verbose: a Boolean: Should we be verbose or not?
n_cores: an integer describing the number of parallel cores that should be used for
running the function
n_per_sample: number of spatial realisations of the occurrence process that we will
simulate for each of the n hyperparameter samples
}
