% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{threshold_occurrence}
\alias{threshold_occurrence}
\title{Function for the threshold occurrence model. This functions takes
a matrix \code{samples} as input, together with a probability \code{nonzero_prob}.
Then, all data points in \code{samples} that are smaller than the \code{nonzero_prob}-quantile
of \code{samples} gets set equal to zero.}
\usage{
threshold_occurrence(samples, nonzero_prob)
}
\description{
Function for the threshold occurrence model. This functions takes
a matrix \code{samples} as input, together with a probability \code{nonzero_prob}.
Then, all data points in \code{samples} that are smaller than the \code{nonzero_prob}-quantile
of \code{samples} gets set equal to zero.
}
